refactor: CustomStringUtils 정의하여 StringAddCalculator Class 분리
